requirements.txt
streamlit
PyPDF2
streamlit-lottie
google-generativeai
